+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add frontend tests.
 
-## [Unreleased] XX-XX-XXXX
+## [1.2.0] 20-09-2024
 
 ### Added